requirements.txt
pybullet==3.2.0
numpy==1.20.3
scipy==1.7.0
matplotlib==3.4.2
roboticstoolbox-python==0.11.0
sympy==1.9
